Indicator order key for graphics

Indicator order key for graphics

indicator_order

Format

A data frame with 148 rows and 12 variables:

SOURCE_TARGET_NAME

character Indicator name (shortened)

PRODUCT

character Indicator name

REPORT_CARD_TITLE

character Indicator name (human-friendly)

Intended.ESP

character Stock name to which the indicator applies

SUBMISSION_YEAR

integer Year the indicator was first included

GATE1_YEAR

integer Year the indicator was added to the ESP

GATE2_YEAR

integer Year the indicator was added to the stock assessment model

REMOVED_YEAR

integer Year the indicator was removed from the ESP

SIGN

integer +1 for positive effect on stock, -1 for negative effect

WEIGHT

integer Relative importance of the indicator for the stock

ORDER

integer Order to display indicators in (no repeated numbers)

ORDER2

integer Order to display indicators in (repeated numbers)

Details

DETAILS