Use files from Google Drive

This function downloads files from Google Drive to a temporary directory to use when rendering an ESP. Intended for internal use by render_esp.

use_google_files(url, ...)

Arguments

url
The URL of the Google folder with the files

...
passed to googledrive::drive_download (ex, overwrite = TRUE)